Auto-generate public/projects.json
