+++ package.json
@@ -4,8 +4,8 @@
   "version": "0.0.0",
   "type": "module",
   "scripts": {
-    "dev": "vite",
-    "build": "tsc && vite build",
+    "dev": "node ./src/generate-projects-json.mjs && vite",
+    "build": "node ./src/generate-projects-json.mjs && tsc && vite build",
     "lint": "biome check",
     "preview": "vite preview"
   },
